Move the main collector tests into `src/collect.rs`.
We now have a new function `all_collectors` which allows
`src/collect.rs` to run these tests on all collectors valid on the
current platform, rather than forcing each collector (e.g.
`perf/collect.rs`) to duplicate these tests.
